refactor: extract key validation from storage/mod.rs into storage/key.rs
